feat: Add cli support
