+++ src/utils.js
@@ -1,12 +1,10 @@
-/*import * as fs from "fs";
+import * as fs from "fs";
 
 const writeJSONFile = (jsonObj) => {
   var file = fs.createWriteStream("tree.json");
-  file.on("error", function (err) {
-  });
+  file.on("error", function (err) {});
   file.write(JSON.stringify(jsonObj));
   file.end();
 };
 
 export { writeJSONFile };
-*/
